pyproject.toml
[project]
name = "gmjj"
version = "0.1.0"
description = "LLM 에이전트 기반 보드게임 룰 안내/TRPG GM 애플리케이션"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langgraph>=1.1.8",
    "openai>=2.29.0",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "supabase>=2.28.3",
    "tavily-python>=0.7.23",
    "uvicorn>=0.42.0",
    "youtube-transcript-api>=1.2.4",
]
